split into chunks
